requirements.txt
asgiref==3.5.2
Django==4.1.3
psycopg2-binary==2.9.5
djangorestframework==3.14.0
sqlparse==0.4.3
Pillow==9.3.0
django-allauth==0.51.0
environs[django]==9.5.0
pytz==2022.6
tzdata==2022.6
